requirements.txt
google-generativeai
feedparser
pygithub
python-dotenv
beautifulsoup4
